Change numCC to uint16_t and increase the max number of ccs

+++ src/sfizz/Defaults.h
@@ -69,7 +69,7 @@ namespace Default
     // Region logic: MIDI conditions
 	constexpr Range<uint8_t> channelRange { 1, 16 };
 	constexpr Range<uint8_t> midiChannelRange { 0, 15 };
-	constexpr Range<uint8_t> ccNumberRange { 0, config::numCCs };
+	constexpr Range<uint16_t> ccNumberRange { 0, config::numCCs };
 	constexpr Range<uint8_t> ccValueRange { 0, 127 };
 	constexpr uint8_t cc { 0 };
 	constexpr Range<int> bendRange { -8192, 8192 };
